Improve stubs for ezodf
