+++ mypy.ini
@@ -40,6 +40,7 @@ warn_unreachable = True
 warn_unused_configs = True
 warn_unused_ignores = True
 ignore_errors = False
+mypy_path = $MYPY_CONFIG_FILE_DIR/src/stubs/
 
 [mypy-dali.abstract_input_filler]
 disallow_any_decorated = False
