+++ src/stubs/ezodf/document.pyi
@@ -1,0 +1,25 @@
+# Copyright 2021 eprbell
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import Optional
+
+from .sheets import Sheets
+
+class PackagedDocument:
+    sheets: Sheets
+
+    def __init__(self, name: str) -> None: ...
+
+def opendoc(filename: str) -> PackagedDocument: ...
+def newdoc(doctype: str, filename: str, template: Optional[str]) -> PackagedDocument: ...
